[APM] E2E Testing: Include E2E resources in operator branch
We plan to leave the resources for testing on a separate branch
for now as we do not want the java code to be in main. Once we
have a centralized place to run the tests, the dependency on
this branch for main can be removed.

The code is copied directly from private-amazon-cloudwatch-staging
on the apm branch.